+++ lensless/admm.py
@@ -1,3 +1,12 @@
+# #############################################################################
+# admm.py
+# =================
+# Authors :
+# Eric BEZZAM [ebezzam@gmail.com]
+# Julien SAHLI [julien.sahli@epfl.ch]
+# #############################################################################
+
+
 import numpy as np
 from lensless.recon import ReconstructionAlgorithm
 from scipy import fft