+++ lensless/gradient_descent.py
@@ -1,3 +1,12 @@
+# #############################################################################
+# gradient_descent.py
+# =================
+# Authors :
+# Eric BEZZAM [ebezzam@gmail.com]
+# Julien SAHLI [julien.sahli@epfl.ch]
+# #############################################################################
+
+
 import numpy as np
 from lensless.recon import ReconstructionAlgorithm
 import inspect