+++ scripts/prep_display_image.py
@@ -55,27 +55,55 @@ import click
     type=int,
     help="Screen resolution in pixels (width, height).",
 )
-def display(fp, pad, output_path, vshift, brightness, screen_res, hshift):
+@click.option(
+    "--rot90",
+    default=0,
+    type=int,
+    help="How many times to rotate provided image by 90 degrees.",
+)
+def display(fp, pad, output_path, vshift, brightness, screen_res, hshift, rot90):
     interpolation = cv2.INTER_NEAREST
 
     # load image
     img_og = cv2.imread(fp, cv2.IMREAD_UNCHANGED)
     img_og = cv2.cvtColor(img_og, cv2.COLOR_BGR2RGB)
 
+    # rotate image
+    if rot90:
+        img_og = np.rot90(img_og, k=rot90)
+
+        # if odd, swap hshift and vshift
+        if rot90 % 2:
+            vshift, hshift = hshift, vshift
+
     if screen_res:
-        img_og_dim = img_og.shape
+        image_height, image_width = img_og.shape[:2]
         img = np.zeros((screen_res[1], screen_res[0], 3), dtype=img_og.dtype)
 
+        # set image with padding and correct aspect ratio
         if screen_res[0] < screen_res[1]:
+
+            max_ratio = screen_res[1] / float(image_height)
+
             new_width = int(screen_res[0] / (1 + 2 * pad / 100))
-            ratio = new_width / float(img_og_dim[0])
-            new_height = int(ratio * img_og_dim[1])
+            ratio = new_width / float(image_width)
+            # new_height = int(ratio * image_height)
+
         else:
+
+            max_ratio = screen_res[0] / float(image_width)
+
             new_height = int(screen_res[1] / (1 + 2 * pad / 100))
-            ratio = new_height / float(img_og_dim[1])
-            new_width = int(ratio * img_og_dim[0])
+            ratio = new_height / float(image_height)
+            # new_width = int(ratio * image_width)
+
+        ratio = min(ratio, max_ratio)
+        new_width = int(ratio * image_width)
+        new_height = int(ratio * image_height)
+
         image_res = (new_width, new_height)
         img_og = cv2.resize(img_og, image_res, interpolation=interpolation)
+
         img[: image_res[1], : image_res[0]] = img_og
 
         # center