+++ scripts/remote_capture.py
@@ -2,165 +2,78 @@
 
 python scripts/remote_capture.py
 
+Check out the `configs/demo.yaml` file for parameters, specifically:
+
+- `rpi`: RPi parameters
+- `capture`: parameters for displaying image
+
 """
 
-
+import hydra
 import os
 import subprocess
-import click
 import cv2
 from pprint import pprint
-import numpy as np
 import matplotlib.pyplot as plt
 import rawpy
 
-from lensless.util import rgb2gray, print_image_info
+
+from lensless.util import rgb2gray, print_image_info, check_username_hostname
 from lensless.plot import plot_image, pixel_histogram
-from lensless.io import load_image
-
-REMOTE_PYTHON = "~/LenslessPiCam/lensless_env/bin/python"
-REMOTE_CAPTURE_FP = "~/LenslessPiCam/scripts/on_device_capture.py"
-SENSOR_MODES = [
-    "off",
-    "auto",
-    "sunlight",
-    "cloudy",
-    "shade",
-    "tungsten",
-    "fluorescent",
-    "incandescent",
-    "flash",
-    "horizon",
-]
+from lensless.io import load_image, save_image
 
 
-@click.command()
-@click.option(
-    "--fn",
-    default="test",
-    type=str,
-    help="File name for recorded image.",
-)
-@click.option(
-    "--username",
-    type=str,
-    default="pi",
-)
-@click.option(
-    "--hostname",
-    type=str,
-    help="Hostname or IP address.",
-)
-@click.option(
-    "--exp",
-    default=0.5,
-    type=float,
-    help="Exposure time in seconds.",
-)
-@click.option(
-    "--iso",
-    default=100,
-    type=int,
-    help="ISO",
-)
-@click.option(
-    "--source",
-    type=click.Choice(["white", "red", "green", "blue"]),
-    default="white",
-    help="Light source.",
-)
-@click.option(
-    "--config_pause",
-    default=2,
-    type=float,
-    help="Pause time for loading / setting camera configuration.",
-)
-@click.option(
-    "--sensor_mode",
-    default="0",
-    type=click.Choice(np.arange(len(SENSOR_MODES)).astype(str)),
-    help="{'off': 0, 'auto': 1, 'sunlight': 2, 'cloudy': 3, 'shade': 4, 'tungsten': 5, "
-    "'fluorescent': 6, 'incandescent': 7, 'flash': 8, 'horizon': 9}",
-)
-@click.option(
-    "--nbits",
-    default=12,
-    type=int,
-    help="Number of bits to set maximum value in histogram. Default is 12 for RPi HQ camera.",
-)
-@click.option(
-    "--legacy",
-    is_flag=True,
-    help="Whether to use legacy camera software, despite being on Bullseye.",
-)
-@click.option(
-    "--rgb",
-    is_flag=True,
-    help="Get RGB data from the Raspberry Pi or reconstruct it here (default). Takes longer to copy RGB data.",
-)
-@click.option(
-    "--gray",
-    is_flag=True,
-    help="Get grayscale data from the Raspberry Pi.",
-)
-@click.option(
-    "--bayer",
-    is_flag=True,
-    help="Whether to save capture data as bayer. Useful for performing color correction later.",
-)
-@click.option(
-    "--gamma",
-    default=None,
-    type=float,
-    help="Gamma factor for plotting.",
-)
-@click.option(
-    "--nbits_out",
-    default=8,
-    type=int,
-    help="Number of bits for output.",
-)
-@click.option("--down", type=float, help="Factor by which to downsample output.", default=None)
-def liveview(
-    fn,
-    username,
-    hostname,
-    exp,
-    iso,
-    config_pause,
-    sensor_mode,
-    nbits,
-    source,
-    rgb,
-    bayer,
-    gamma,
-    nbits_out,
-    legacy,
-    gray,
-    down,
-):
-    if bayer:
-        assert not rgb
-        assert not gray
+@hydra.main(version_base=None, config_path="../configs", config_name="demo")
+def liveview(config):
+
+    bayer = config.capture.bayer
+    rgb = config.capture.rgb
+    gray = config.capture.gray
+
+    username, hostname = check_username_hostname(config.rpi.username, config.rpi.hostname)
+    legacy = config.capture.legacy
+    nbits_out = config.capture.nbits_out
+    fn = config.capture.raw_data_fn
+    gamma = config.capture.gamma
+    source = config.capture.source
+    plot = config.plot
+
+    if config.save:
+        if config.output is not None:
+            # make sure output directory exists
+            os.makedirs(config.output, exist_ok=True)
+            save = config.output
+        else:
+            save = os.getcwd()
+    else:
+        save = False
+
+    # proceed with capture
+    # if bayer:
+    #     assert not rgb
+    #     assert not gray
     assert hostname is not None
 
     # take picture
     remote_fn = "remote_capture"
     print("\nTaking picture...")
     pic_command = (
-        f"{REMOTE_PYTHON} {REMOTE_CAPTURE_FP} --fn {remote_fn} --exp {exp} --iso {iso} "
-        f"--config_pause {config_pause} --sensor_mode {sensor_mode} --nbits_out {nbits_out}"
+        f"{config.rpi.python} {config.capture.script} bayer={bayer} fn={remote_fn} exp={config.capture.exp} iso={config.capture.iso} "
+        f"config_pause={config.capture.config_pause} sensor_mode={config.capture.sensor_mode} nbits_out={config.capture.nbits_out}"
     )
-    if nbits > 8:
-        pic_command += " --sixteen"
-    if rgb:
-        pic_command += " --rgb"
-    if legacy:
-        pic_command += " --legacy"
-    if gray:
-        pic_command += " --gray"
-    if down:
-        pic_command += f" --down {down}"
+    if config.capture.nbits > 8:
+        pic_command += " sixteen=True"
+    if config.capture.rgb:
+        pic_command += " rgb=True"
+    if config.capture.legacy:
+        pic_command += " legacy=True"
+    if config.capture.gray:
+        pic_command += " gray=True"
+    if config.capture.down:
+        pic_command += f" down={config.capture.down}"
+    if config.capture.awb_gains:
+        pic_command += f" awb_gains=[{config.capture.awb_gains[0]},{config.capture.awb_gains[1]}]"
+
     print(f"COMMAND : {pic_command}")
     ssh = subprocess.Popen(
         ["ssh", "%s@%s" % (username, hostname), pic_command],
@@ -240,6 +153,8 @@ def liveview(
         # more pythonic? https://stackoverflow.com/questions/250283/how-to-scp-in-python
         remotefile = f"~/{remote_fn}.png"
         localfile = f"{fn}.png"
+        if save:
+            localfile = os.path.join(save, localfile)
         print(f"\nCopying over picture as {localfile}...")
         os.system('scp "%s@%s:%s" %s' % (username, hostname, remotefile, localfile))
 
@@ -247,20 +162,34 @@ def liveview(
             img = load_image(localfile, verbose=True)
 
         else:
-            # get white balance gains
-            if bayer:
-                red_gain = 1
-                blue_gain = 1
+
+            if not bayer:
+                red_gain = config.camera.red_gain
+                blue_gain = config.camera.blue_gain
             else:
-                red_gain = float(result_dict["Red gain"])
-                blue_gain = float(result_dict["Blue gain"])
+                red_gain = None
+                blue_gain = None
+            # # get white balance gains
+            # if red_gain is None:
+            #     red_gain = float(result_dict["Red gain"])
+            # if blue_gain is None:
+            #     blue_gain = float(result_dict["Blue gain"])
+
+            # # get white balance gains
+            # if bayer:
+            #     red_gain = 1
+            #     blue_gain = 1
+            # else:
+            #     red_gain = float(result_dict["Red gain"])
+            #     blue_gain = float(result_dict["Blue gain"])
 
             # load image
             print("\nLoading picture...")
+
             img = load_image(
                 localfile,
                 verbose=True,
-                bayer=True,
+                bayer=bayer,
                 blue_gain=blue_gain,
                 red_gain=red_gain,
                 nbits_out=nbits_out,
@@ -270,35 +199,55 @@ def liveview(
             if not bayer:
                 cv2.imwrite(localfile, cv2.cvtColor(img, cv2.COLOR_RGB2BGR))
 
+    # save image as viewable 8 bit
+    fp = os.path.join(save, f"{fn}_8bit.png")
+    save_image(img, fp)
+
     # plot RGB
-    if not gray:
-        ax = plot_image(img, gamma=gamma)
-        ax.set_title("RGB")
+    if plot:
+        if not gray:
+            ax = plot_image(img, gamma=gamma)
+            ax.set_title("RGB")
+            if save:
+                plt.savefig(os.path.join(save, f"{fn}_plot.png"))
 
-        # plot red channel
-        if source == "red":
-            img_1chan = img[:, :, 0]
-        elif source == "green":
-            img_1chan = img[:, :, 1]
-        elif source == "blue":
-            img_1chan = img[:, :, 2]
+            # plot red channel
+            if source == "red":
+                img_1chan = img[:, :, 0]
+            elif source == "green":
+                img_1chan = img[:, :, 1]
+            elif source == "blue":
+                img_1chan = img[:, :, 2]
+            else:
+                img_1chan = rgb2gray(img[None, :, :, :])
+            ax = plot_image(img_1chan)
+            if source == "white":
+                ax.set_title("Gray scale")
+            else:
+                ax.set_title(f"{source} channel")
+            if save:
+                plt.savefig(os.path.join(save, f"{fn}_1chan.png"))
+
+            # plot histogram, useful for checking clipping
+            pixel_histogram(img)
+            if save:
+                plt.savefig(os.path.join(save, f"{fn}_hist.png"))
+            pixel_histogram(img_1chan)
+            if save:
+                plt.savefig(os.path.join(save, f"{fn}_1chan_hist.png"))
+
         else:
-            img_1chan = rgb2gray(img[None, :, :, :])
-        ax = plot_image(img_1chan)
-        if source == "white":
-            ax.set_title("Gray scale")
-        else:
-            ax.set_title(f"{source} channel")
+            ax = plot_image(img, gamma=gamma)
+            if save:
+                plt.savefig(os.path.join(save, f"{fn}_plot.png"))
+            pixel_histogram(img)
+            if save:
+                plt.savefig(os.path.join(save, f"{fn}_hist.png"))
 
-        # plot histogram, useful for checking clipping
-        pixel_histogram(img)
-        pixel_histogram(img_1chan)
+        plt.show()
 
-    else:
-        ax = plot_image(img, gamma=gamma)
-        pixel_histogram(img)
-
-    plt.show()
+    if save:
+        print(f"\nSaved plots to: {save}")
 
 
 if __name__ == "__main__":