+++ CHANGELOG.rst
@@ -14,12 +14,16 @@ Added
 ~~~~~
 
 - Sensor module.
+- Single-script and Telegram demo.
+- Link and citation for JOSS.
+- Authors at top of source code files.
+- Add paramiko as dependency for remote capture and display.
 
 
 Changed
 ~~~~~~~
 
-- 
+- Simpler remote capture and display scripts with Hydra.
 
 
 Bugfix